requirements.txt
torch>=1.9.0
transformers>=4.21.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0
faiss-cpu>=1.7.0,<2.0.0
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<1.26.0
scikit-learn>=1.1.0,<1.4.0
spacy>=3.4.0,<4.0.0
